src/modules/bot/src/requirements.txt
async-generator==1.10
attrs==21.4.0
black==22.6.0
certifi==2022.6.15
cffi==1.15.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
h11==0.13.0
idna==3.3
mypy-extensions==0.4.3
outcome==1.2.0
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
selenium==4.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
